+++ EssentialFeed/EssentialFeedTests/Helpers/SharedTestHelpers.swift
@@ -1,0 +1,16 @@
+//
+//  SharedTestHelpers.swift
+//  EssentialFeedTests
+//
+//  Created by Nayem, Mufakkharul | Nil | GSSD on 2025/06/12.
+//
+
+import Foundation
+
+func anyURL() -> URL {
+    return URL(string: "https://any-url.com")!
+}
+
+func anyNSError() -> NSError {
+    return NSError(domain: "any error", code: 0)
+}
